requirements.txt
networkx==1.10
mock==1.0.1
coverage==3.7.1
pyyaml>=3.11
GitPython>=2.0.1